updated the express app flow, only have one middle for the `graphql` endpoint
