+++ controllers/index.js
@@ -1,0 +1,1 @@
+// empty file. :) for controllers
