--- index.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <meta http-equiv="X-UA-Compatible" content="ie=edge">
-  <title>Document</title>
-</head>
-<body>
-  hello world HTML File
-</body>
-
-<script>
- var body = JSON.stringify({query: "{ getBuildings { building_type } }"})
-
- fetch('http://localhost:3000/graphql', {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: body
-  })
-  .then(res => res.json())
-  .then(data => console.log(data));
-
-</script>
-</html>
